Introduction

This tiny package provides a simple framework for fetching data from the Yahoo Finance API. It is designed to be used with DataFrames.jl and YFinance.jl. However, the DataFrames.jl dependency can be included as an extension in Julia 1.9+. The package is designed for my personal use and is not intended to be used by others. However, if you find it useful, feel free to use it.This tiny package provides a simple framework for fetching data from the Yahoo Finance API. It is designed to be used with DataFrames.jl and YFinance.jl. However, the DataFrames.jl dependency can be included as an extension in Julia 1.9+. The package is designed for my personal use and is not intended to be used by others. However, if you find it useful, feel free to use it.

Quick Start

The package can be installed using the Julia package manager. From the Julia REPL, type ] to enter the Pkg REPL mode and run:

pkg> add https://github.com/shayandavoodii/MFF.git

Fetch Close Prices for ["AAPL", "MSFT"] from 2020-01-01 to 2020-01-10 and store the result in a DataFrame:

julia> using MFF

julia> get_data(Val(:df), ["AAPL", "MSFT"], "2020-01-01", "2020-01-10")
6×3 DataFrame
 Row │ date        AAPL     MSFT    
     │ Date        Float64  Float64 
─────┼──────────────────────────────
   1 │ 2020-01-02  73.4494  155.762
   2 │ 2020-01-03  72.7353  153.822
   3 │ 2020-01-04  73.3149  154.22
   4 │ 2020-01-05  72.9701  152.814
   5 │ 2020-01-06  74.1439  155.248
   6 │ 2020-01-07  75.7188  157.187

API

MFF.get_data — Method

get_data(::Val{:df}, stock::String, startdt::String, enddt::String; prprty::String="adjclose", rng::Nothing=nothing, fixdt::Bool=true)::DataFrame

Fetch data from Yahoo Finance and return a DataFrame.

Arguments

::Val{:df}: Return a DataFrame.

stock::String: The stock ticker.

startdt::String: The start date. The format is "YYYY-MM-DD".

enddt::String: The end date. The format is "YYYY-MM-DD".

prprty::String="adjclose": The property to fetch. The other options are "open", "high", "low", "close", "vol", "timestamp".

rng::Nothing=nothing: The range of the data. The other options are "1d", "5d", "1mo", "3mo", "6mo", "1y", "2y", "5y", "10y".

fixdt::Bool=true: Fix the dates. If true, the dates will be fixed to the range of the data. If false, the dates will be the dates of the data.

Returns

::DataFrame: The DataFrame of the data.

Methods

get_data(::Val{:df}, stock::String, startdt::String, enddt::String; prprty::String="adjclose", rng::Nothing=nothing, fixdt::Bool=true)::DataFrame

get_data(::Val{:vec}, stock::String, startdt::String, enddt::String; prprty::String="adjclose", rng::Nothing=nothing)::Vector{Vector}

Examples

julia> using MFF

julia> get_data(Val(:df), "AAPL", "2020-01-10", "2020-01-15", fixdt=false)
3×2 DataFrame
 Row │ date        AAPL
     │ Date        Float64
─────┼─────────────────────
   1 │ 2020-01-10  75.89
   2 │ 2020-01-13  77.5113
   3 │ 2020-01-14  76.4646

julia> get_data(Val(:df), "AAPL", "2020-01-10", "2020-01-15")
3×2 DataFrame
 Row │ date        AAPL
     │ Date        Float64
─────┼─────────────────────
   1 │ 2020-01-10  75.89
   2 │ 2020-01-11  77.5113
   3 │ 2020-01-12  76.4646