API

MFF.fix_dates! — Method

fix_dates!(df::DataFrame, dates::Vector{Date})

Fix date related column in df inplace by changing its frequency to daily.

Arguments

df::DataFrame: DataFrame object.

dates::Vector{Date}: Vector of dates to be used as the new date column.

Example

julia> df = DataFrame(col1=[1, 2, 3, 4, 5], col2=["a", "b", "c", "d", "e"])
5×2 DataFrame
 Row │ col1   col2
     │ Int64  String
─────┼───────────────
   1 │     1  a
   2 │     2  b
   3 │     3  c
   4 │     4  d
   5 │     5  e

julia> dates = collect(Date("2020-01-10"):Day(2):Date("2020-01-19"))
5-element Vector{Date}:
2020-01-10
2020-01-12
2020-01-14
2020-01-16
2020-01-18

julia> fix_dates!(df, dates);

julia> df
5×3 DataFrame
 Row │ date        col1   col2
     │ Date        Int64  String
─────┼───────────────────────────
   1 │ 2020-01-10      1  a
   2 │ 2020-01-11      2  b
   3 │ 2020-01-12      3  c
   4 │ 2020-01-13      4  d
   5 │ 2020-01-14      5  e

MFF.get_data — Method

get_data(::Val{:df}, stock::String, startdt::String, enddt::String; prprty::String="adjclose", rng::Nothing=nothing, fixdt::Bool=true)::DataFrame

Fetch data from Yahoo Finance and return a DataFrame.

Arguments

::Val{:df}: Return a DataFrame.

stock::String: The stock ticker.

startdt::String: The start date. The format is "YYYY-MM-DD".

enddt::String: The end date. The format is "YYYY-MM-DD".

prprty::String="adjclose": The property to fetch. The other options are "open", "high", "low", "close", "vol", "timestamp".

rng::Nothing=nothing: The range of the data. The other options are "1d", "5d", "1mo", "3mo", "6mo", "1y", "2y", "5y", "10y".

fixdt::Bool=true: Fix the dates. If true, the dates will be fixed to the range of the data. If false, the dates will be the dates of the data.

plot::Bool=false: Plot the data.

Returns

::DataFrame: The DataFrame of the data.

Methods

get_data(::Val{:df}, stock::String, startdt::String, enddt::String; prprty::String="adjclose", rng::Nothing=nothing, fixdt::Bool=true)::DataFrame

get_data(::Val{:vec}, stock::String, startdt::String, enddt::String; prprty::String="adjclose", rng::Nothing=nothing)::Vector{Vector}

Examples

julia> using MFF

julia> get_data(Val(:df), "AAPL", "2020-01-10", "2020-01-15", fixdt=false)
3×2 DataFrame
 Row │ date        AAPL
     │ Date        Float64
─────┼─────────────────────
   1 │ 2020-01-10  75.89
   2 │ 2020-01-13  77.5113
   3 │ 2020-01-14  76.4646

julia> get_data(Val(:df), "AAPL", "2020-01-10", "2020-01-15")
3×2 DataFrame
 Row │ date        AAPL
     │ Date        Float64
─────┼─────────────────────
   1 │ 2020-01-10  75.89
   2 │ 2020-01-11  77.5113
   3 │ 2020-01-12  76.4646

julia> get_data(Val(:df), "AAPL", "2020-01-10", "2020-01-15", prprty="open")
3×2 DataFrame
 Row │ date        AAPL
     │ Date        Float64
─────┼─────────────────────
   1 │ 2020-01-10  75.956
   2 │ 2020-01-11  76.2103
   3 │ 2020-01-12  77.4477

julia> get_data(Val(:df), ["AAPL", "MSFT"], "2020-01-10", "2020-01-15", prprty="high")
3×3 DataFrame
Row │ date        AAPL     MSFT
    │ Date        Float64  Float64
─────┼──────────────────────────────
  1 │ 2020-01-10  76.4622  158.283
  2 │ 2020-01-11  77.5382  158.37
  3 │ 2020-01-12  77.6605  158.652

julia> get_data(Val(:vec), ["AAPL", "MSFT"], "2020-01-10", "2020-01-15", prprty="high")
3×2 Matrix{Float64}:
76.4622  158.283
77.5382  158.37
77.6605  158.652